Move absence queries to a separate file
